ecudata: initial version of the script
The inital version of the script. Running the script will
open a file dialog to search for csv files.

Currently the supported format is the one that VIPEC produces,
which is of following type:

<Name>, <PC Datalog - date>
<empty line>
<comma separated headers>
<comma separated units>
<empty line>
<data>

the first row is timestamp in seconds. i recommend using section
time when exporting the logs, this way the log starts from 0 and
its much easier to fit into the composition.

Signed-off-by: Janne Kotka <lekonna@gmail.com>